requirements.txt
# Core
pydantic>=2.6
python-dotenv>=1.0

# UI
streamlit>=1.32

# Agent framework (used in Step 4)
langgraph>=0.0.40
langchain>=0.1.16

# LLM client (Step 4)
openai>=1.12
